refactor: 코드 리뷰 반영 필드명 자세히 변경, Nickname 클래스 util 패키지로 분리, ERD 이미지 수정
