+++ domain/src/main/java/com/health/entity/MealEntity.java
@@ -40,7 +40,7 @@ public class MealEntity extends BaseEntity{
     @Embedded
     @AttributeOverrides({
             @AttributeOverride(name = "kCal", column = @Column(name = "kcal")),
-            @AttributeOverride(name = "carbohydrate", column = @Column(name = "carbo")),
+            @AttributeOverride(name = "carbohydrate", column = @Column(name = "carbohydrate")),
             @AttributeOverride(name = "protein", column = @Column(name = "protein")),
             @AttributeOverride(name = "fat", column = @Column(name = "fat"))
     })
